modified:   CMakeLists.txt 	new file:   hello.cpp 	modified:   include/loader/loader.h 	modified:   src/loader/loader.cpp 	new file:   vendor/adder/CMakeLists.txt 	new file:   vendor/adder/include/adder/adder.h 	new file:   vendor/adder/src/adder/adder.cpp

+++ cmake/include/loader/loader.h
@@ -1,8 +1,7 @@
 #ifndef __LOADER_H__
 #define __LOADER_H__
-
 namespace loader {
-double load();
+int load();
 } // namespace loader
 
 #endif//__LOADER_H__
